Libc: Add times and getimeofday and implement syscall

+++ src/libraries/libc/sys/types.h
@@ -16,7 +16,7 @@ typedef int tid_t;
 typedef int uid_t;
 typedef int gid_t;
 typedef int sid_t;
-typedef unsigned long sigset_t;
+typedef unsigned int sigset_t;
 typedef int idtype_t;
 typedef int id_t;
 typedef long long time_t;
